Improving the build.
This commit changes are as follows:
1. build.sh - sometimes the file launch.sh needs executable mode change permission so now its happening automatically inside the script.
2. launch.sh - Checks if a docker container (instantiation) exists if it does it runs it otherwise it creates it from scratch and then runs it.
another change here is the removal of --rm which before would remove the container and all of the usr/local installed files including open3d.h files which on terminal termination a new open3d build was necessary. So now that's not the case and open3d installation can happen only once.
3. Dockerfile - Following new build issues, we install python3.10 and as a result downgrade numpy (for numpy_eigen).