Minor tweaks - to get the job done on my personal laptop using WSL.
This fixes a last minute issue with the DecompUtil package that weren't downloaded successfully.
But most likely the problem was due to using the command:
sed -i 's/cmake_minimum_required(VERSION 2.8.3)/cmake_minimum_required(VERSION 3.5)/' src/decomp_util/CMakeLists.txt
when the CMakeLists.txt file didn't exist for a reason.
Commenting out probably was the the solution that fixed the issue.